Show planned times with the tile closed
